- Исправил расчёт и хранение суммы заказа (BigDecimal, правильный тип в БД). - Сделал легковесный DTO для заказов (id, сумма, дата). - Реализовал подгрузку активных заказов в профиле.

+++ src/main/resources/db/migration/V32__Ititial_table_orders.sql
@@ -1,6 +1,6 @@
 CREATE TABLE orders (
     id BIGSERIAL PRIMARY KEY,
-    total_price DOUBLE PRECISION NOT NULL,
+    total_price NUMERIC(12,2) NOT NULL,
     date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     first_name VARCHAR(255) NOT NULL,
     last_name VARCHAR(255) NOT NULL,
